fix resource provider order
